backend/requirements.txt
django==3.1.2
djangorestframework==3.11.1
django-cors-headers==3.5.0
drf-yasg==1.17.1
drf_writable_nested==0.5.1
djangorestframework_simplejwt==4.3.0
rest-condition==1.0.3
